% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{addVector}
\alias{addVector}
\alias{addVector,Dense2SparseHDF5-method}
\title{addVector runs wilcox test on the columns of the sparse matrix}
\usage{
addVector(dat, offset, alloc)
}
\arguments{
\item{dat}{the dense matrix}

\item{alloc}{the amount of data that has to fit into the object}
}
\value{
null
}
\description{
count the not zeros in a dense matrix
}
