\name{anRpackage-package}
\alias{anRpackage-package}
\alias{anRpackage}
\docType{package}
\title{\packageTitle{anRpackage}}
\description{\packageDescription{anRpackage}}
\section{Package Content}{\packageIndices{anRpackage}}
\author{\packageAuthor{anRpackage}}
\section{Maintainer}{\packageMaintainer{anRpackage}}
\keyword{package}
