% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.class.Dense2SparseHDF5.R
\docType{class}
\name{Dense2SparseHDF5}
\alias{Dense2SparseHDF5}
\title{Class a simple interface to store a sparse Matrix and fill it with a HUGE dense one
from an HDF5 file connection.}
\description{
Class a simple interface to store a sparse Matrix and fill it with a HUGE dense one
from an HDF5 file connection.

Class a simple interface to store a sparse Matrix and fill it with a HUGE dense one
from an HDF5 file connection.
}
\details{
The main intension here is to keep memory footprint as low as possible 
while also trying to be as fast as possible.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Matrix}}{the sparse matrix}

\item{\code{file}}{the HDF5 connection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{Dense2SparseHDF5$print()}}
\item \href{#method-new}{\code{Dense2SparseHDF5$new()}}
\item \href{#method-toSparse}{\code{Dense2SparseHDF5$toSparse()}}
\item \href{#method-toSparseVector}{\code{Dense2SparseHDF5$toSparseVector()}}
\item \href{#method-clone}{\code{Dense2SparseHDF5$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
print a summary of the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dense2SparseHDF5$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
initialize the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dense2SparseHDF5$new(fname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fname}}{the hdf5 file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toSparse"></a>}}
\subsection{Method \code{toSparse()}}{
convert a dense matrix entry in the hdf5 file to sparse
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dense2SparseHDF5$toSparse(slot = "matrix", block_size = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slot}}{the slot to convert (default 'matrix')}

\item{\code{block_size}}{the read block size ( default 1000)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toSparseVector"></a>}}
\subsection{Method \code{toSparseVector()}}{
convert a dense matrix entry in the hdf5 file to sparse
This function does not directly modify the sparse matrix oin c++,
but converts the matrix into a i j and x vector using c++ and the R Matrix Package.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dense2SparseHDF5$toSparseVector(slot = "matrix", block_size = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{slot}}{the slot to convert (default 'matrix')}

\item{\code{block_size}}{the read block size ( default 1000)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dense2SparseHDF5$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
