% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Matr}
\alias{Matr}
\alias{Matr,Dense2SparseHDF5-method}
\title{Matr will create an R saprse matrix resized to the correct number of entries, but not filled}
\usage{
Matr(nrow, ncol, size)
}
\arguments{
\item{nrow}{row count}

\item{ncol}{col count}

\item{size}{the amount of not zero entries}
}
\value{
the propper resized dense matrix (hopefully)
}
\description{
count the not zeros in a dense matrix
}
