% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NotNull}
\alias{NotNull}
\alias{NotNull,Dense2SparseHDF5-method}
\title{NotNull returns the amount of not zwero entries in the matrix/vector}
\usage{
NotNull(X)
}
\arguments{
\item{X}{a dense vector}
}
\value{
the amount of not zero entries
}
\description{
count the not zeros in a dense matrix
}
