% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add}
\alias{add}
\alias{add,Dense2SparseHDF5-method}
\title{add runs wilcox test on the columns of the sparse matrix}
\usage{
add(X, dat, offset, alloc, progressBar = TRUE)
}
\arguments{
\item{X}{the sparse matrix}

\item{dat}{the dense matrix}

\item{alloc}{the amount of data that has to fit into the return object}
}
\value{
null
}
\description{
count the not zeros in a dense matrix
}
